Enables registering concrete richlets instead of their class names
In that case, the ZK will get the concrete class and it does not have to
construct the new instance.

So, we'd be able to set the richlet's properties before registering it
to the ZK. Now, we have to set all properties in the richlet's
constructor what is not adequate in some cases.